data/katia-executive-grade-analysis-v1.md: v0.9.55 (frozen)
KATIA-v1 SERIE, DEVELOPMENT FROZEN IN v0.9.55

Since the "WHEN TO RAG? Always!" suggestion has been implemented also in
Google Gemini, all the previous interactions (chat session informations)
are provided to the next prompt in read-only mode as if that prompt were
the first one. On the other hand, caching information between-prompts is
not an accessible feature anymore.

Under this perspective [CSC] is provided by a sort of RAG about all the
previous interactions while the rules on how to use it are still working.
While the concept of IBPS (persistency between prompts) fall apart, soon
after it started to work as expected. Which brought commands like
'show-savings' and 'print-ops' to show 'none'.

Questioning the AI about these failures is almost useless because the
explanations are created as post-hoc artifacts (aka hallucinations).
In fact, these explanations are not consistent in terms of checking.
The AI fails "smoothly" to execute some part of the prompt and it has
no idea of the root cause of those failures.

Hence, development is broken in its essence, especially because without
persistence between prompts, debugging becomes extremely difficult.
In fact, in some terms every prompt is the first prompt but with more RAG.
So, debugging what? A sort of Alzheimer for which short-memory is gone?

-> data/katia-executive-grade-analysis-v1.txt: v0.9.55

Signed-off-by: Roberto A. Foglietta <roberto.foglietta@gmail.com>